pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "networkx>=3.4.2",
    "notebook>=7.4.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "snowflake>=1.2.0",
    "snowflake-connector-python>=3.14.0",
    "streamlit>=1.44.1",
]
